+++ util-crypto/ed25519/index.ts
@@ -2,7 +2,6 @@
 /**
  * @summary Implements ed25519 operations
  */
-export { convertPublicKeyToCurve25519, convertSecretKeyToCurve25519 } from './convertKey.ts';
 export { ed25519DeriveHard } from './deriveHard.ts';
 export { ed25519PairFromRandom } from './pair/fromRandom.ts';
 export { ed25519PairFromSecret } from './pair/fromSecret.ts';