+++ util/is/bigInt.ts
@@ -8,7 +8,7 @@
  * <BR>
  *
  * ```javascript
- * import { isBigInt } from 'https://deno.land/x/polkadot@0.2.31/util/mod.ts';
+ * import { isBigInt } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * console.log('isBigInt', isBigInt(123_456n)); // => true
  * ```