+++ x-textdecoder/fallback.ts
@@ -1,6 +1,6 @@
 
 export class TextDecoder {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars,no-useless-constructor
+  // eslint-disable-next-line no-useless-constructor
   constructor (_?: 'utf-8' | 'utf8') {
     // nothing
   }