+++ x-textdecoder/packageInfo.ts
@@ -1,3 +1,3 @@
 
 
-export const packageInfo = { name: '@polkadot/x-textdecoder', path: new URL(import.meta.url).pathname, type: 'deno', version: '0.2.31' };
+export const packageInfo = { name: '@polkadot/x-textdecoder', path: new URL(import.meta.url).pathname, type: 'deno', version: '11.1.1' };